Crate brontes

Brontes

This is the main binary crate for the Brontes project. It uses several other crates in the workspace, which are documented separately:

brontes-core: Tracing for the Brontes project.

brontes-inspect: Mev Inspectors for MEV detection.

brontes-types: Defines the main types used across Brontes.

brontes-classifier: Contains the classifier logic pertaining to transaction trace classification & normalization.

brontes-metrics: Handles metrics collection and reporting.

brontes-pricing: Handles DEX pricing

brontes-database: Handles database related functionalities.

brontes-database-libmdbx: Provides libmdbx database functionalities for our local libmdbx database.

reth-tracing-ext: Provides extended tracing capabilities to match transaction traces to their corresponding logs.

Please refer to the individual crate documentation for more details.

Structs

Brontes

DataBatching

Constants

PROMETHEUS_ENDPOINT_IP

PROMETHEUS_ENDPOINT_PORT